Feat: regenerate helheim mocks (#18)
**Tools: add hel as a tool**

We use hel for generating mocks. Not importing hel as a tool means that
we may use different versions of hel whenenever we go to regenerate
mocks. We should align on a version, so this puts it in the go.mod.

Should just be able to run `go run git.sr.ht/~nelsam/hel/v3` in the
relevant package dirs to regenerate mocks from now on.

**Tests: regenerated helheim mocks**

* Had to manually replace the package name with `loggregator_v2` in both
  instances of regeneration
* Removed the unused bits of the generated code to make golangci-lint
  happy

**Fix some golangci-lint warnings**

* Remove some unused code
* Ignore warnings we don't care about