Example app: make it a separate go module
This example app should not share the same go.mod as the production
code. It should be maintained separately.

Since we probably won't import that app anywhere, it's module name
doesn't matter, so I named it `example`.

Had to bump the version of github.com/cactus/go-statsd-client imported
to v5 because it was so old that no version had been released yet, and
there was no statsd package to import.